feat: Add page for others' profiles
